% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineup.R
\name{buildLineUp}
\alias{buildLineUp}
\title{lineup - factory for LineUp HTMLWidget}
\usage{
buildLineUp(
  x,
  width = "100\%",
  height = NULL,
  elementId = NULL,
  dependencies = crosstalk::crosstalkLibs()
)
}
\arguments{
\item{x}{LineUpBuilder object}

\item{width}{width of the element}

\item{height}{height of the element}

\item{elementId}{unique element id}

\item{dependencies}{include crosstalk dependencies}
}
\value{
html lineup widget
}
\description{
lineup - factory for LineUp HTMLWidget
}
