% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineup.R
\docType{data}
\name{.lineupDefaultOptions}
\alias{.lineupDefaultOptions}
\title{LineUpjs module}
\format{An object of class \code{list} of length 15.}
\usage{
.lineupDefaultOptions
}
\description{
a htmlwidget wrapper around LineUpJS (\url{https://lineup.js.org})
}
\keyword{datasets}
