% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineup.R
\name{.lineupImpl}
\alias{.lineupImpl}
\title{LineUpjs module}
\usage{
.lineupImpl(data, width, height, elementId, options, ranking, dependencies,
  lineupType, ...)
}
\description{
a htmlwidget wrapper around LineUpJS (\url{https://lineup.js.org})
}
