% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineup.R
\name{lineup-shiny}
\alias{lineup-shiny}
\alias{lineupOutput}
\alias{renderLineup}
\title{Shiny bindings for lineup}
\usage{
lineupOutput(outputId, width = "100\%", height = "800px")

renderLineup(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'800px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{expr}{An expression that generates a lineup}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Output and render functions for using lineup within Shiny
applications and interactive Rmd documents.
}
